refactor: log 도메인 분리
- API 엔드포인트(자원)을 기준으로 3개의 서브 도메인으로 분리
  - elapsed-time
  - traffic
  - success-rate

- 테스트 케이스 리팩터링 완료
