+++ backend/console-server/src/log/elapsed-time/dto/get-top5-elapsed-time.dto.ts
@@ -2,7 +2,7 @@ import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
 import { IsNumber } from 'class-validator';
 
-export class GetSpeedRankDto {
+export class GetTop5ElapsedTimeDto {
     @IsNumber()
     @Type(() => Number)
     @ApiProperty({
