+++ backend/console-server/src/log/elapsed-time/metric/host-avg-elapsed-time.metric.ts
@@ -1,7 +1,7 @@
 import { IsNumber, IsString } from 'class-validator';
 import { Type } from 'class-transformer';
 
-export class SpeedRankMetric {
+export class HostAvgElapsedTimeMetric {
     @IsString()
     host: string;
