+++ backend/console-server/src/log/elapsed-time/metric/path-elapsed-time.metric.ts
@@ -1,6 +1,6 @@
 import { IsString } from 'class-validator';
 
-export class ElapsedTimeByPathMetric {
+export class PathElapsedTimeMetric {
     @IsString()
     avg_elapsed_time: string;
