+++ backend/console-server/src/log/log.module.ts
@@ -1,14 +1,14 @@
 import { Module } from '@nestjs/common';
-import { LogService } from './log.service';
 import { LogController } from './log.controller';
-import { LogRepository } from './log.repository';
-import { Clickhouse } from '../clickhouse/clickhouse';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Project } from '../project/entities/project.entity';
+import { ClickhouseModule } from '../clickhouse/clickhouse.module';
+import { LogService } from './log.service';
+import { LogRepository } from './log.repository';
 
 @Module({
-    imports: [TypeOrmModule.forFeature([Project])],
+    imports: [TypeOrmModule.forFeature([Project]), ClickhouseModule],
+    providers: [LogService, LogRepository],
     controllers: [LogController],
-    providers: [LogService, LogRepository, Clickhouse],
 })
 export class LogModule {}
