+++ backend/console-server/src/log/success-rate/dto/get-project-success-rate-response.dto.ts
@@ -1,7 +1,7 @@
 import { Expose, Type } from 'class-transformer';
 import { ApiProperty } from '@nestjs/swagger';
 
-export class GetSuccessRateByProjectResponseDto {
+export class GetProjectSuccessRateResponseDto {
     @ApiProperty({
         description: '프로젝트의 이름',
         example: 'watchducks',
