+++ backend/console-server/src/log/success-rate/dto/get-project-success-rate.dto.ts
@@ -1,7 +1,7 @@
 import { IsString } from 'class-validator';
 import { ApiProperty } from '@nestjs/swagger';
 
-export class GetSuccessRateByProjectDto {
+export class GetProjectSuccessRateDto {
     @IsString()
     @ApiProperty({
         description: '프로젝트 이름',
