+++ backend/console-server/src/log/success-rate/metric/error-rate.metric.ts
