+++ backend/console-server/src/log/traffic/dto/get-traffic-daily-difference-response.dto.ts
