+++ backend/console-server/src/log/traffic/dto/get-traffic-top5.dto.ts
@@ -2,7 +2,7 @@ import { ApiProperty } from '@nestjs/swagger';
 import { Expose, Type } from 'class-transformer';
 import { IsNotEmpty, IsNumber } from 'class-validator';
 
-export class GetTrafficRankDto {
+export class GetTrafficTop5Dto {
     @ApiProperty({
         example: 9,
         description: '기수',
